# PERF: get_dummies